Graphics
This project is inspired by WorldQuant University's Applied Data Science Lab course. It is being hosted on DataWars with their permission. Please refer to their course for more details.  
In this project, you'll work with a dataset of homes for sale in Brazil. You'll need to perform different Data Analysis tasks to discover if there are regional differences in the real estate market. Also, you will look at southern Brazil region to see if there is a relationship between home size and price, similar to what you saw with housing in some states in Mexico.